Add to_bin and from_bin serialization/deserialization
One of the key advantages of xor filters is they are small and
immutable. This means sending them over the wire or storing them on
disk or in a database is a very useful usecase. This patch adds support
for serializing/deserializing xor8 and xor16 filters to and from an
Erlang binary to make storing/transferring xor filters easy.