Rework resourse usage and memory management
The xor filter does not require the value list to exist once the xor
filter is populated. With the tracking of that removed, there was no
need for the intermediate struct. Thus, the filter_resource structs have
been removed and the enif resources point directly to xor(8|16)_t
structs. Additionally, since the xor filters, once constructed, are not
of a large size (a xor16 with 10 million keys is only 165 megabytes) the
explicit free() functions are of questionable value and add a lot of
complexity when we can just let the reference count go to 0 and have the
data get garbage collected by the BEAM.