Prevent undefined behavior in exor_t
If an exor_t is created but never finalized, it triggers UB when
garbage collected. The problem is that the gc resouce callback frees
`buffer`. Sometimes `buffer` may be zero by chance only, in which case
the destructor does not attempt to free it. But if those uninitialized
bytes are anything else the descrutor will attempt to free an invalid
address and likely segfault. Regardless, this is all UB, segfault or
not, and the fix is to initialize the memory.